django/requirements.txt
asgiref==3.8.1
django==5.1.3
gunicorn==23.0.0
sqlparse==0.4.4
Pillow==11.1.0
psycopg2==2.9.10
djangorestframework==3.15.2